refactor: TicketDataAdapter 인터페이스 ui레이어로 이동
