+++ app/src/main/java/woowacourse/movie/data/local/adapter/TicketAdapter.kt
@@ -3,8 +3,8 @@ package woowacourse.movie.data.local.adapter
 import android.content.Context
 import woowacourse.movie.data.local.database.MovieDatabase.Companion.getMovieDatabase
 import woowacourse.movie.data.local.entity.TicketEntity
-import woowacourse.movie.domain.data.TicketDataAdapter
 import woowacourse.movie.domain.ticket.Ticket
+import woowacourse.movie.ui.view.data.TicketDataAdapter
 
 class TicketAdapter(private val context: Context) : TicketDataAdapter {
     private val dao by lazy { getMovieDatabase(context).ticketDao() }
